# docs/openapi.yaml
openapi: 3.1.0
info:
  title: psp-sandbox
  version: 0.3.0
  summary: A fake payment provider for tests and CI that fails the way real providers do.
  description: |
    Two APIs on one port. The provider API (`/v1`) is what your application calls.
    The control API (`/_sandbox`) is what your tests call to inspect and drive the
    sandbox. Callbacks to your application are described under `webhooks`.

    Prose documentation with the details: docs/api.md, docs/scenarios.md and
    docs/callbacks.md in the repository.
  license:
    name: MIT
    identifier: MIT
...
servers:
  - url: http://localhost:8090
...
paths:
  /v1/payments:
    post:
      tags: [provider]
      operationId: createPayment
      summary: Create a payment
      description: |
        The payment is created as `pending` and moves on asynchronously, with a
        callback per status change, unless the scenario says otherwise. The scenario
        comes from the `X-Sandbox-Scenario` header, else from the rules file, else
        `PSP_DEFAULT_SCENARIO`.
      security:
        - bearer: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Scenario'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePayment'
      responses:
        '201':
          description: Created. May be held or never sent, depending on the scenario.
          headers:
            Idempotent-Replayed:
              $ref: '#/components/headers/IdempotentReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        5XX:
          $ref: '#/components/responses/ServerError'
    get:
      tags: [provider]
      operationId: listPayments
      summary: List payments, oldest first
      security:
        - bearer: []
      parameters:
        - name: reference
          in: query
          required: false
          description: Only payments with exactly this reference.
          schema:
            type: string
      responses:
        '200':
          description: Payments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentList'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/payments/{id}:
    get:
      tags: [provider]
      operationId: getPayment
      summary: Get a payment
      security:
        - bearer: []
      parameters:
        - $ref: '#/components/parameters/PaymentId'
      responses:
        '200':
          description: The payment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Any key is accepted unless `PSP_API_KEY` is set.
  parameters:
    PaymentId:
      name: id
      in: path
      required: true
      schema:
        type: string
        examples: [pay_01J9Z3K8Q2W5]
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Same key and same request return the stored answer. Same key with another
        request, or while the first is running, is `409 idempotency_conflict`. Error
        answers are not stored.
      schema:
        type: string
    Scenario:
      name: X-Sandbox-Scenario
      in: header
      required: false
      description: Scenario for this payment, `name; param=value`. See `GET /_sandbox/scenarios`.
      schema:
        type: string
        examples: ['duplicate_callback; times=3; parallel=true']
  headers:
    IdempotentReplayed:
      description: '`true` when the answer was stored for an earlier request with the same key.'
      schema:
        type: string
        const: 'true'
  responses:
    InvalidRequest:
      description: Invalid body, parameter or scenario (`invalid_request`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or wrong API key (`unauthorized`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Unknown id (`not_found`).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: '`invalid_state` or `idempotency_conflict`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: |
        `server_error` from the `server_error_then_success` scenario, nothing was
        created; or `internal_error`, a bug in the sandbox.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Amount:
      description: Integer in minor units.
      type: integer
      format: int64
      minimum: 0
    Currency:
      type: string
      pattern: '^[A-Z][A-Z0-9]{2,4}$'
      examples: [EUR, JPY, USDT]
    Time:
      type: string
      format: date-time
    CreatePayment:
      type: object
      additionalProperties: false
      required: [amount, currency]
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1
        currency:
          $ref: '#/components/schemas/Currency'
        reference:
          description: Your id. Not required to be unique.
          type: string
        capture:
          type: string
          enum: [auto, manual]
          default: auto
        callback_url:
          type: string
          format: uri
        return_url:
          description: Where the 3DS page sends the customer after they act.
          type: string
          format: uri
        metadata:
          type: object
          additionalProperties:
            type: string
    PaymentStatus:
      type: string
      enum: [pending, requires_action, authorized, captured, partially_refunded, refunded, failed, canceled, disputed, chargeback_lost, chargeback_won]
    Payment:
      type: object
      required: [id, status, amount, captured_amount, refunded_amount, currency, capture, scenario, created_at, updated_at]
      properties:
        id:
          type: string
          examples: [pay_01J9Z3K8Q2W5]
        status:
          $ref: '#/components/schemas/PaymentStatus'
        amount:
          $ref: '#/components/schemas/Amount'
        captured_amount:
          $ref: '#/components/schemas/Amount'
        refunded_amount:
          $ref: '#/components/schemas/Amount'
        currency:
          $ref: '#/components/schemas/Currency'
        reference:
          description: Left out when empty.
          type: string
        capture:
          type: string
          enum: [auto, manual]
        scenario:
          type: string
        failure_reason:
          description: Only on failed payments.
          type: string
          enum: [insufficient_funds, do_not_honor, expired_card, fraud_suspected, generic_decline, lost_card, stolen_card, incorrect_cvc, processing_error, authentication_failed]
        action_url:
          description: Only while the payment is `requires_action`. The page where the customer authenticates.
          type: string
          format: uri
        created_at:
          $ref: '#/components/schemas/Time'
        updated_at:
          $ref: '#/components/schemas/Time'
        metadata:
          description: Left out when empty.
          type: object
          additionalProperties:
            type: string
    PaymentList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_request, unauthorized, not_found, method_not_allowed, idempotency_conflict, invalid_state, clock_not_manual, amount_exceeds_captured, server_error, internal_error]
            message:
              type: string